Starting AWS download process
Looking at ways to access the GEFS data through AWS. eco4cast has some data, but that seems to be for only certain sites. Wanting to look through to see how to get the forecasts for individual days and then how to subset for the area that we are interested in.