# 00_NOAA_GEFS_functions.R is my attempt at changing the default GEFS pulls (0.5deg) to instead pull teh 0.25 deg data. I do not know how successful it was. I was practicing with point data for now.
gefs_datapull_test.R seems to be working fairly reliably now. Have updated to include a section where we can specify our own point. We can also make a bounding box, according to Carl.  This script outputs 3hrly forecast data. We will need to think about how we are aggregating the data to a daily level for each of the variables of interest.